[LoopAccessAnalysis] Keep results even when not all pointers were analysed
Currently if there's any pointer that loop access analysis couldn't
analyse then all of the results are discarded. This patch changes that
so they're not, and we generate the runtime check information for
those pointers that we could analyse, as transformations may still be
able to make use of the partial information.

Of the transformations that use LoopAccessAnalysis, only
LoopVersioningLICM changes behaviour as a result of this change. This
is because the others either:
 * Check canVectorizeMemory, which will return false when we have
   partial pointer information as analyzeLoop() will return false.
 * Examine the dependencies returned by getDepChecker(), which will be
   empty as we exit analyzeLoop if we have partial pointer information
   before calling areDepsSafe(), which is what fills in the dependency
   information.